Fixed bugs with store and node.blockchain not syncing w/ demo functions
It was possible for contents from old blockchain variations to survive a (store) call if the contents of chaindata0 were fewer than the contents of backup_chaindata0. This was noticed during testing, and now the directory is cleared before the contents are copied over. 

Also there was an issue with (consensus) and (mine) functions where if these were called the blockchain value of that node would not be updated in the demo file. This has been fixed.